# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OwnBrief API
  description: |
    OwnBrief는 YouTube 플레이리스트를 기반으로 AI가 생성한 팟캐스트를 제공하는 서비스입니다.
    
    이 API는 사용자 인증, 팟캐스트 생성 및 관리, YouTube 플레이리스트 통합, 
    온보딩 프로세스 등의 기능을 제공합니다.
    
    ## 인증
    대부분의 엔드포인트는 NextAuth를 통한 세션 인증이 필요합니다.
    
    ## 기술 스택
    - Next.js 14 (App Router)
    - Prisma ORM
    - PostgreSQL
    - Google Gemini AI
    - Google OAuth & YouTube API
    - Supabase Storage
  version: 1.0.0
  contact:
    name: OwnBrief Team
servers:
  - url: http://localhost:3000
    description: 로컬 개발 서버
  - url: https://your-production-url.vercel.app
    description: 프로덕션 서버
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth 세션 쿠키
...
paths:
...
  /api/user/delete:
    delete:
      tags:
        - User
      summary: 사용자 계정 삭제
      description: 사용자 계정 및 관련 모든 데이터 삭제
      security:
        - SessionAuth: []
      responses:
        '200':
          description: 계정 삭제 성공
        '401':
          description: 인증 필요

  # ==================== Onboarding ====================
...